Moved functionality from headers to c files so that we can compile
